+++ src/App.jsx
@@ -1,0 +1,24 @@
+import "./App.css";
+// eslint-disable-next-line import/extensions
+import logo from "./logo.svg";
+
+const App = () => (
+  <div className="App">
+    <header className="App-header">
+      <img alt="logo" className="App-logo" src={logo} />
+      <p>
+        Edit <code>src/App.js</code> and save to reload.
+      </p>
+      <a
+        className="App-link"
+        href="https://reactjs.org"
+        rel="noopener noreferrer"
+        target="_blank"
+      >
+        Learn React
+      </a>
+    </header>
+  </div>
+);
+
+export default App;
