+++ src/App.test.jsx
@@ -1,0 +1,9 @@
+import { render, screen } from "@testing-library/react";
+
+import App from "./App";
+
+test("renders learn react link", () => {
+  render(<App />);
+  const linkElement = screen.getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
+});
